# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Sanctum offers a simple-to-use authentication system like Laravel Sanctum.",
        "title": "Go Sanctum",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "This endpoint login a user with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "This endpoint login a user with email and password",
                "parameters": [
                    {
                        "description": "User Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllerv1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseCreatedUser"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseError"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "This endpoint gets a user with its token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "This endpoint gets a user with its token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication Bearer header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseGetUser"
                        }
                    },
                    "401": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint creates a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "This endpoint creates a new user",
                "parameters": [
                    {
                        "description": "User Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllerv1.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseCreatedUser"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseError"
                        }
                    },
                    "500": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllerv1.CreateRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "email@email.com"
                },
                "name": {
                    "type": "string",
                    "example": "name"
                },
                "password": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
        "controllerv1.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "email@email.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
        "swagger.ResponseCreatedUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.createdUser"
                }
            }
        },
        "swagger.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "swagger.ResponseGetUser": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.user"
                }
            }
        },
        "swagger.createdUser": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "1|V96CSH7oe3Pjr174sL4s24rCcnjZhWWw0IBUUjeh"
                },
                "user": {
                    "$ref": "#/definitions/swagger.user"
                }
            }
        },
        "swagger.user": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "email@email.com"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "name"
                }
            }
        }
    }
}